refactor: extract get ballot contract logic into seperate function
